use attach inline from lookup
