+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "aquality-tracking-ui",
-  "version": "1.0.1",
+  "version": "1.0.2",
   "scripts": {
     "ng": "ng",
     "start": "ng serve",
